Minimum failure case with 32-bit dma_addr ops
There are two writes in gasket_page_table.c that are required to be
64-bit. If these values are switched to writeq, the device will work
(even though all the other reads/writes from gasket_core are 32-bit).

This setup writes as two 32-bit numbers and reads them back. Switching
to the writeq option will allow the device to work on the Dev Board.